encoding/xml: fix printing of namespace prefix in tag names
Prefix displays in XML when defined in tag names.
The URL of the name space is also returned for an End Token as
documentation requires to improve idempotency of Marshal/Unmarshal.

Translating the prefix is popping the NS which was unavailable
for translation. Translate for an End Token has been moved inside
the pop element part.

Fixes #9519

Change-Id: Id7a14d07c106a76a487b5c4e28d9d563fe061c60